[ALPS07253642] vcodec: fix coverity
ALPS07155102 10779701 CERT-C Characters and Strings
ALPS07155196 10779906 CERT-C Characters and Strings
ALPS07158825 11148405 CERT-C Characters and Strings
ALPS07253637 11333332 Negative array index read
ALPS07159183 11301087 Invalid type in argument to printf format specifier
ALPS07159328 11301238 Invalid type in argument to printf format specifier
ALPS07159365 11301268 Invalid type in argument to printf format specifier
ALPS07159446 11301346 Invalid type in argument to printf format specifier
ALPS07253642 11333337 Unsigned compared against 0

MTK-Commit-Id: 8a67ac6cf2c6ab179768312ef7d2d84791f42044

CR-Id: ALPS07253642
Feature: Video Driver

+++ drivers/misc/mediatek/vcp/rv/vcp_excep.c
@@ -446,7 +446,7 @@ static unsigned int vcp_crash_dump(enum vcp_core_id id)
 static void vcp_prepare_aed_dump(char *aed_str, enum vcp_core_id id)
 {
 	char *vcp_A_log = NULL;
-	size_t offset = 0;
+	int offset = 0;
 
 	pr_debug("[VCP] %s begins:%s\n", __func__, aed_str);
 	vcp_dump_last_regs(mmup_enable_count());
